Reworking Drill Screens to be Componentized (#140)
* DrillList has been componentized for main drill list and profile view to be sectioned off

* Moved glyphs to be in section header instead of drill card

* Drlll Description componentized for description screen and description modal and deleted image carousel

* ran yarn pretty

* slight optimization to getUnique

* Removed prettyDrillType and modified drillType

---------
