+++ components/drillCard.js
@@ -1,38 +1,14 @@
 import { Link } from "expo-router";
 import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
-import { getCombinedDrillTitle } from "~/Utility";
 
 function DrillCard(props) {
-  // console.log(props);
   return (
     <Link href={props.hrefString} asChild>
       <TouchableOpacity style={styles.cardContainer}>
         <View style={styles.cardContent}>
-          <Text style={styles.title}>{getCombinedDrillTitle(props.drill)}</Text>
+          <Text style={styles.title}>{props.drill.subType}</Text>
           <View style={styles.specContainer}>
-            <Text style={styles.inputText}>
-              {props.drill["inputs"]
-                .map((input) => {
-                  let retVal = "";
-                  switch (input.id) {
-                    case "carry":
-                      retVal = "↑";
-                      break;
-                    case "sideLanding":
-                      retVal = "↔︎";
-                      break;
-                    case "strokes":
-                      retVal = "#";
-                      break;
-                    default:
-                      retVal = "?";
-                  }
-                  return retVal;
-                })
-                .join(" ") +
-                " x" +
-                props.drill["reps"]}
-            </Text>
+            <Text style={styles.inputText}>{"x" + props.drill["reps"]}</Text>
           </View>
         </View>
       </TouchableOpacity>
@@ -56,7 +32,6 @@ const styles = StyleSheet.create({
   },
   title: {
     fontSize: 16,
-    fontWeight: "bold",
     color: "#333",
   },
   specContainer: {
